condensed select field for both item forms into one component
